Change test file to use .tsx
